Revert "Merge branch 'main' into database"
This reverts commit 27669a503574c34f61095789467a807db6a87c19, reversing
changes made to 853babc082294f47d555d23a365c5096b1a8d84b.
