+++ components/Calendar.js
@@ -1,0 +1,20 @@
+'use client';
+import { Calendar, momentLocalizer } from "react-big-calendar";
+import moment from "moment";
+import "react-big-calendar/lib/css/react-big-calendar.css";
+// import 'react-big-calendar/lib/sass/styles';
+
+
+const localizer = momentLocalizer(moment);
+
+export default function MyCalendar({ events }) {
+  return (
+    <Calendar
+      localizer={localizer}
+      events={events}
+      startAccessor="start"
+      endAccessor="end"
+      style={{ height: 500 }}
+    />
+  );
+}
