+++ src/app/globals.css
@@ -20,42 +20,3 @@ body {
   font-family: Arial, Helvetica, sans-serif;
 }
 
-.rbc-calendar {
-  display: flex;
-  flex-direction: column;
-  max-width: 800px;
-  background: #fff;
-  border-radius: 5px;
-  margin: 0 auto;
-}
-
-.date-time-container {
-  display: flex;
-  margin: 0px auto;
-  margin-top: 12px;
-  margin-bottom: 5px;
-  max-width: 800px;
-}
-
-.add-button-container {
-  display: flex;
-  justify-content: right;
-  margin: 0px auto;
-  max-width: 800px;
-}
-
-.height-400 {
-  height: 400px;
-}
-
-@media (max-width: 768px) {
-  .date-time-container {
-    margin-top: 75px;
-  }
-}
-
-
-
-
-
-
