--- src/app/providers.jsx
@@ -1,11 +1,0 @@
-'use client'
-
-import {NextUIProvider} from '@nextui-org/react'
-
-export function Providers({children}) {
-  return (
-    <NextUIProvider>
-      {children}
-    </NextUIProvider>
-  )
-}
